image-service/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
Pillow>=10.2.0
qrcode>=7.4.2
numpy>=1.24.4  # Compatible with Python 3.8
python-multipart>=0.0.9
pydantic>=2.5.3
scipy>=1.10.1  # Last version supporting Python 3.8
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.1  # Last version supporting Python 3.8